[O2B-1343] Extract GAQ detectors logic to separate service and controller (#1712)
* add method to service

* update GAQ dets in ML sync

* fix test

* docs

* add service

* add test

* cleanup

* add controller

* add router

* cleanup

* adjust fornt

* add api test

* add

* cleanup

* fix ML sync

* add test

* add docs

* ref

* docs

* cleanup

* fix

* fix

* fix

* fix

* fix

* expose router

* fix

* rename

* fix

* rename, docs
