+++ lib/public/views/QcFlags/GaqFlags/GaqFlagsOverviewModel.js
@@ -96,7 +96,7 @@ export class GaqFlagsOverviewModel extends OverviewPageModel {
         this._gaqDetectors$.setCurrent(RemoteData.loading());
         try {
             const { data: gaqDetectors } =
-                await getRemoteData(`/api/dataPasses/gaqDetectors?dataPassId=${this._dataPassId}&runNumber=${this._runNumber}`);
+                await getRemoteData(`/api/gaqDetectors?dataPassId=${this._dataPassId}&runNumber=${this._runNumber}`);
             this._gaqDetectors$.setCurrent(RemoteData.success(gaqDetectors));
         } catch (error) {
             this._gaqDetectors$.setCurrent(RemoteData.failure(error));
@@ -140,7 +140,7 @@ export class GaqFlagsOverviewModel extends OverviewPageModel {
         this._setGaqDetectorsRequestResult.setCurrent(RemoteData.loading());
 
         try {
-            await jsonPost('/api/dataPasses/gaqDetectors', {
+            await jsonPost('/api/gaqDetectors', {
                 dataPassId: this._dataPassId,
                 runNumbers: [this._runNumber],
                 dplDetectorIds: this._gaqDetectorPickerModel.selected,
