+++ lib/server/controllers/dataPasses.controller.js
@@ -55,71 +55,6 @@ const listDataPassesHandler = async (req, res) => {
     }
 };
 
-// eslint-disable-next-line valid-jsdoc
-/**
- *
- */
-const setGaqDetectors = async (req, res) => {
-    const validatedDTO = await dtoValidator(
-        DtoFactory.bodyOnly(Joi.object({
-            dataPassId: Joi.number(),
-            runNumbers: Joi.array().items(Joi.number()).unique(),
-            dplDetectorIds: Joi.array().items(Joi.number()).unique(),
-        })),
-        req,
-        res,
-    );
-    if (validatedDTO) {
-        try {
-            const {
-                dataPassId,
-                runNumbers,
-                dplDetectorIds,
-            } = validatedDTO.body;
-
-            const data = await dataPassService.setGaqDetectors(
-                dataPassId,
-                runNumbers,
-                dplDetectorIds,
-            );
-
-            res.status(201).json({ data });
-        } catch (error) {
-            updateExpressResponseFromNativeError(res, error);
-        }
-    }
-};
-
-// eslint-disable-next-line valid-jsdoc
-/**
- * List DPL detectors contributing to GAQ for given data pass and run
- */
-const listGaqDetectors = async (req, res) => {
-    const validatedDTO = await dtoValidator(
-        DtoFactory.queryOnly(Joi.object({
-            dataPassId: Joi.number().required(),
-            runNumber: Joi.number().required(),
-        })),
-        req,
-        res,
-    );
-    if (validatedDTO) {
-        try {
-            const {
-                dataPassId,
-                runNumber,
-            } = validatedDTO.query;
-
-            const data = await dataPassService.getGaqDetectors(dataPassId, runNumber);
-            res.json({ data });
-        } catch (error) {
-            updateExpressResponseFromNativeError(res, error);
-        }
-    }
-};
-
 exports.DataPassesController = {
     listDataPassesHandler,
-    setGaqDetectors,
-    listGaqDetectors,
 };
