+++ lib/server/monalisa-synchronization/MonAlisaSynchronizer.js
@@ -21,7 +21,7 @@ const { extractLhcPeriod } = require('../utilities/extractLhcPeriod.js');
 const { getOrCreateLhcPeriod } = require('../services/lhcPeriod/getOrCreateLhcPeriod.js');
 const { Op } = require('sequelize');
 const { RunDefinition } = require('../../domain/enums/RunDefinition.js');
-const { dataPassService } = require('../services/dataPasses/DataPassService.js');
+const { gaqDetectorService } = require('../services/gaq/GaqDetectorsService.js');
 const { SkimmingStage } = require('../../domain/enums/SkimmingStage.js');
 
 const SKIMMING_DATA_PASS_NAME_REGEX = /skimming/;
@@ -118,7 +118,7 @@ class MonAlisaSynchronizer {
 
             if (newRunsToBeAssociated.length > 0) {
                 await dataPass.addRuns(newRunsToBeAssociated);
-                await dataPassService.useDefaultGaqDetectors(
+                await gaqDetectorService.useDefaultGaqDetectors(
                     dataPass.id,
                     newRunsToBeAssociated.map(({ runNumber }) => runNumber),
                     { transaction },
