+++ lib/server/routers/dataPasses.router.js
@@ -11,29 +11,10 @@
  *  or submit itself to any jurisdiction.
  */
 
-const { BkpRoles } = require('../../domain/enums/BkpRoles.js');
 const { DataPassesController } = require('../controllers/dataPasses.controller.js');
-const { rbacMiddleware } = require('../middleware/rbac.middleware.js');
 
 exports.dataPassesRouter = {
     path: '/dataPasses',
-    children: [
-        {
-            method: 'get',
-            controller: DataPassesController.listDataPassesHandler,
-        },
-        {
-            path: 'gaqDetectors',
-            children: [
-                {
-                    method: 'post',
-                    controller: [rbacMiddleware([BkpRoles.ADMIN, BkpRoles.GAQ]), DataPassesController.setGaqDetectors],
-                },
-                {
-                    method: 'get',
-                    controller: DataPassesController.listGaqDetectors,
-                },
-            ],
-        },
-    ],
+    method: 'get',
+    controller: DataPassesController.listDataPassesHandler,
 };
