+++ lib/server/routers/index.js
@@ -20,6 +20,7 @@ const { dplProcessRouter } = require('./dplProcess.router.js');
 const environmentRoute = require('./environments.router');
 const { eosReportRouter } = require('./eosReport.router.js');
 const flpsRoute = require('./flps.router');
+const { gaqDetectorsRouter } = require('./gaqDetectors.router.js');
 const lhcFillRoute = require('./lhcFills.router');
 const logRoute = require('./logs.router');
 const runTypesRoute = require('./runTypes.router');
@@ -47,6 +48,7 @@ const routes = [
     environmentRoute,
     eosReportRouter,
     flpsRoute,
+    gaqDetectorsRouter,
     lhcFillRoute,
     lhcFillRoute,
     lhcPeriodsRouter,
