+++ test/api/dataPasses.test.js
@@ -15,8 +15,6 @@ const { expect } = require('chai');
 const request = require('supertest');
 const { server } = require('../../lib/application');
 const { resetDatabaseContent } = require('../utilities/resetDatabaseContent.js');
-const { DetectorType } = require('../../lib/domain/enums/DetectorTypes');
-const { BkpRoles } = require('../../lib/domain/enums/BkpRoles');
 
 const LHC22b_apass1 = {
     id: 1,
@@ -294,51 +292,6 @@ module.exports = () => {
                     expect(titleError.detail).to.equal('"query.page.limit" must be greater than or equal to 1');
                     done();
                 });
-        });
-    });
-
-    describe('POST /api/dataPasses/gaqDetectors', () => {
-        it('should successfully set GAQ detectors', async () => {
-            const dataPassId = 3;
-            const runNumbers = [49, 56];
-            const detectorIds = [4, 7];
-            const response = await request(server).post(`/api/dataPasses/gaqDetectors?token=${BkpRoles.GAQ}`).send({
-                dataPassId,
-                runNumbers,
-                dplDetectorIds: detectorIds,
-            });
-            expect(response.status).to.be.equal(201);
-            expect(response.body.data).to.have.all.deep.members(runNumbers
-                .flatMap((runNumber) => detectorIds.map((detectorId) => ({ dataPassId, runNumber, detectorId }))));
-        });
-
-        it('should fail to set GAQ detectors because of insufficient permission', async () => {
-            const dataPassId = 3;
-            const runNumbers = [49, 56];
-            const detectorIds = [4, 7];
-            const response = await request(server).post(`/api/dataPasses/gaqDetectors?token=${BkpRoles.GUEST}`).send({
-                dataPassId,
-                runNumbers,
-                dplDetectorIds: detectorIds,
-            });
-            expect(response.status).to.be.equal(403);
-
-            const { errors } = response.body;
-            expect(errors.find(({ title }) => title === 'Access denied')).to.not.be.null;
-        });
-    });
-
-    describe('GET /api/dataPasses/gaqDetectors', () => {
-        it('should return 200 with the list of GAQ detectors', async () => {
-            const response = await request(server).get('/api/dataPasses/gaqDetectors?dataPassId=3&runNumber=56');
-
-            expect(response.status).to.equal(200);
-            const { data } = response.body;
-            expect(data).to.be.an('array');
-            expect(data).to.have.all.deep.members([
-                { id: 7, name: 'FT0', type: DetectorType.PHYSICAL },
-                { id: 4, name: 'ITS', type: DetectorType.PHYSICAL },
-            ]);
         });
     });
 };
