+++ test/lib/server/services/dataPasses/DataPassService.test.js
@@ -15,17 +15,7 @@ const { expect } = require('chai');
 const { resetDatabaseContent } = require('../../../../utilities/resetDatabaseContent.js');
 const assert = require('assert');
 const { NotFoundError } = require('../../../../../lib/server/errors/NotFoundError.js');
-const {
-    dataPassService,
-    DEFAULT_GAQ_DETECTORS_FOR_PROTON_PROTON_RUNS,
-    DEFAULT_GAQ_DETECTORS_FOR_LEAD_LEAD_RUNS,
-} = require('../../../../../lib/server/services/dataPasses/DataPassService.js');
-const { BadParameterError } = require('../../../../../lib/server/errors/BadParameterError.js');
-const { DetectorType } = require('../../../../../lib/domain/enums/DetectorTypes.js');
-const { runService } = require('../../../../../lib/server/services/run/RunService.js');
-const DataPassRepository = require('../../../../../lib/database/repositories/DataPassRepository.js');
-const RunRepository = require('../../../../../lib/database/repositories/RunRepository.js');
-const { Op } = require('sequelize');
+const { dataPassService } = require('../../../../../lib/server/services/dataPasses/DataPassService.js');
 
 const LHC22b_apass1 = {
     id: 1,
@@ -181,72 +171,6 @@ module.exports = () => {
             const { rows: dataPasses } = await dataPassService.getAll(dto.query);
             expect(dataPasses.map(({ name }) => name)).to.have
                 .ordered.members(['LHC22a_apass1', 'LHC22a_apass2_skimmed', 'LHC22a_skimming', 'LHC22b_apass1', 'LHC22b_apass2']);
-        });
-    });
-    describe('Manage GAQ detectors', () => {
-        const dataPassId = 3;
-        it('should successfully set GAQ detectors', async () => {
-            const runNumbers = [49, 56];
-            const detectorIds = [4, 7];
-            const data = await dataPassService.setGaqDetectors(dataPassId, runNumbers, detectorIds);
-            expect(data).to.be.have.all.deep.members(runNumbers
-                .flatMap((runNumber) => detectorIds.map((detectorId) => ({ dataPassId, runNumber, detectorId }))));
-        });
-        it('should fail to set GAQ detectors because of missing association', async () => {
-            let errorMessage = `No association between data pass with id ${dataPassId} and following runs: 1`;
-            assert.rejects(
-                () => dataPassService.setGaqDetectors(dataPassId, [1], [4]),
-                new BadParameterError(errorMessage),
-            );
-            errorMessage = `No association between runs and detectors: ${JSON.stringify([[56, 'CPV']])}`;
-            assert.rejects(
-                () => dataPassService.setGaqDetectors(dataPassId, [105, 56], [1]),
-                new BadParameterError(errorMessage),
-            );
-        });
-
-        it('should get GAQ detectors', async () => {
-            const detectors = await dataPassService.getGaqDetectors(3, 56);
-            expect(detectors).to.be.an('array');
-            expect(detectors).to.be.lengthOf(2);
-            expect(detectors).to.have.all.deep.members([
-                { id: 7, name: 'FT0', type: DetectorType.PHYSICAL },
-                { id: 4, name: 'ITS', type: DetectorType.PHYSICAL },
-            ]);
-        });
-
-        it('should successfully set default GAQ detectors', async () => {
-            /**
-             * Default GAQ detectors for runs with given pdpBeamType
-             *      pp: ['TPC', 'ITS', 'FT0']
-             *      PbPb: ['TPC', 'ITS', 'FT0', 'ZDC']
-             */
-
-            const newRuns = [
-                { runNumber: 777770, pdpBeamType: 'pp', detectors: ['CPV', ...DEFAULT_GAQ_DETECTORS_FOR_PROTON_PROTON_RUNS].join(',') },
-                { runNumber: 777771, pdpBeamType: 'pp', detectors: ['CPV', ...DEFAULT_GAQ_DETECTORS_FOR_PROTON_PROTON_RUNS.slice(1)].join(',') },
-                { runNumber: 888880, pdpBeamType: 'PbPb', detectors: ['CPV', ...DEFAULT_GAQ_DETECTORS_FOR_LEAD_LEAD_RUNS].join(',') },
-                { runNumber: 888881, pdpBeamType: 'PbPb', detectors: ['CPV', ...DEFAULT_GAQ_DETECTORS_FOR_LEAD_LEAD_RUNS.slice(1)].join(',') },
-            ];
-            for (const runData of newRuns) {
-                await runService.create(runData);
-            }
-            const dataPassId = 3;
-            const dataPass = await DataPassRepository.findOne({ where: { id: dataPassId } });
-            const runNumbers = newRuns.map(({ runNumber }) => runNumber);
-
-            await dataPass.addRuns(await RunRepository.findAll({ where: { runNumber: { [Op.in]: runNumbers } } }));
-
-            await dataPassService.useDefaultGaqDetectors(dataPassId, runNumbers);
-
-            expect((await dataPassService.getGaqDetectors(dataPassId, 777770)).map(({ name }) => name)).to
-                .have.all.members(DEFAULT_GAQ_DETECTORS_FOR_PROTON_PROTON_RUNS);
-            expect((await dataPassService.getGaqDetectors(dataPassId, 777771)).map(({ name }) => name)).to
-                .have.all.members(DEFAULT_GAQ_DETECTORS_FOR_PROTON_PROTON_RUNS.slice(1));
-            expect((await dataPassService.getGaqDetectors(dataPassId, 888880)).map(({ name }) => name)).to
-                .have.all.members(DEFAULT_GAQ_DETECTORS_FOR_LEAD_LEAD_RUNS);
-            expect((await dataPassService.getGaqDetectors(dataPassId, 888881)).map(({ name }) => name)).to
-                .have.all.members(DEFAULT_GAQ_DETECTORS_FOR_LEAD_LEAD_RUNS.slice(1));
         });
     });
 };
