+++ test/lib/server/services/index.js
@@ -31,6 +31,7 @@ const UserSuite = require('./user/index.js');
 const SimulationPassesSuite = require('./simulationPasses/index.js');
 const QcFlagsSuite = require('./qualityControlFlag/index.js');
 const TriggerCountersSuite = require('./triggerCounters/index.js');
+const GaqDetectorSuite = require('./gaq');
 
 module.exports = () => {
     before(resetDatabaseContent);
@@ -43,6 +44,7 @@ module.exports = () => {
     describe('Environment history item', EnvironmentHistoryItemSuite);
     describe('EOS report', EosReportSuite);
     describe('Flp role', FlpRoleSuite);
+    describe('GaqDetector', GaqDetectorSuite);
     describe('LHC fill suite', LhcFillSuite);
     describe('Logs', LogSuite);
     describe('RunType', RunTypeSuite);
